fix(coderd,site): gate chat advisor and virtual desktop behind experiments, delete experiments page
The Coder Agents feature created an 'Experiments' page that conflated
deployment-level feature flags with runtime DB toggles. The page contained
advisor, virtual desktop, computer use provider, and debug logging admin
gate settings, none of which were gated behind any experiment.

This change:
- Adds ExperimentChatAdvisor and ExperimentChatVirtualDesktop to
  codersdk/deployment.go
- Gates the advisor and desktop/computer-use-provider API routes behind
  RequireExperimentWithDevBypass middleware
- Passes experiments to chatd.Server and checks them before attaching the
  advisor tool or enabling desktop, so stale DB values cannot bypass the
  experiment gate
- Deletes the AgentSettingsExperimentsPage and its route, sidebar nav item,
  and sidebar view entry
- Moves the AdvisorSettings and VirtualDesktopSettings components to the
  Agents settings page, visible only when the corresponding experiment is
  enabled
- Moves the AdminChatDebugLoggingSettings component to the Lifecycle page
  (debug logging has a deployment flag, not an experiment)
- Updates docs to reflect the new experiment-gated structure
- Updates test server creation to include ExperimentsKnown by default
