+++ coderd/x/chatd/chatd.go
@@ -190,6 +190,7 @@ type Server struct {
 
 	aibridgeTransportFactory *atomic.Pointer[aibridge.TransportFactory]
 	aiGatewayRoutingEnabled  bool
+	experiments              codersdk.Experiments
 
 	// Configuration
 	pendingChatAcquireInterval time.Duration
@@ -3166,6 +3167,7 @@ type Config struct {
 	Clock                          quartz.Clock
 	AIBridgeTransportFactory       *atomic.Pointer[aibridge.TransportFactory]
 	AIGatewayRoutingEnabled        bool
+	Experiments                    codersdk.Experiments
 
 	PrometheusRegistry prometheus.Registerer
 
@@ -3262,6 +3264,7 @@ func New(ps pubsub.Pubsub, cfg Config) *Server {
 		},
 		aibridgeTransportFactory:   cfg.AIBridgeTransportFactory,
 		aiGatewayRoutingEnabled:    cfg.AIGatewayRoutingEnabled,
+		experiments:                cfg.Experiments,
 		pendingChatAcquireInterval: pendingChatAcquireInterval,
 		maxChatsPerAcquire:         maxChatsPerAcquire,
 		inFlightChatStaleAfter:     inFlightChatStaleAfter,
