+++ site/src/api/typesGenerated.ts
@@ -4570,6 +4570,8 @@ export type Experiment =
 	| "ai-gateway-cost-control"
 	| "agent-app-tabs"
 	| "auto-fill-parameters"
+	| "chat-advisor"
+	| "chat-virtual-desktop"
 	| "example"
 	| "mcp-server-http"
 	| "minimum-implicit-member"
@@ -4583,6 +4585,8 @@ export const Experiments: Experiment[] = [
 	"ai-gateway-cost-control",
 	"agent-app-tabs",
 	"auto-fill-parameters",
+	"chat-advisor",
+	"chat-virtual-desktop",
 	"example",
 	"mcp-server-http",
 	"minimum-implicit-member",
