+++ site/src/pages/AISettingsPage/CoderAgentsPage/CoderAgentsPage.tsx
@@ -7,12 +7,19 @@ import {
 } from "react-query";
 import { API } from "#/api/api";
 import {
+	chatAdvisorConfig,
+	chatComputerUseProvider,
+	chatDesktopEnabled,
 	chatModelConfigs,
 	chatPersonalModelOverridesAdminSettings,
+	updateChatAdvisorConfig,
+	updateChatComputerUseProvider,
+	updateChatDesktopEnabled,
 	updateChatPersonalModelOverridesAdminSettings,
 } from "#/api/queries/chats";
 import type * as TypesGen from "#/api/typesGenerated";
 import { useAuthenticated } from "#/hooks/useAuthenticated";
+import { useDashboard } from "#/modules/dashboard/useDashboard";
 import { RequirePermission } from "#/modules/permissions/RequirePermission";
 import { pageTitle } from "#/utils/page";
 import { CoderAgentsPageView } from "./CoderAgentsPageView";
@@ -48,8 +55,13 @@ const updateChatModelOverrideMutation = (
 
 const CoderAgentsPage: FC = () => {
 	const { permissions } = useAuthenticated();
+	const { experiments } = useDashboard();
 	const queryClient = useQueryClient();
 	const canEditDeploymentConfig = permissions.editDeploymentConfig;
+	const showAdvisorSettings = experiments.includes("chat-advisor");
+	const showVirtualDesktopSettings = experiments.includes(
+		"chat-virtual-desktop",
+	);
 
 	const personalModelOverridesAdminSettingsQuery = useQuery({
 		...chatPersonalModelOverridesAdminSettings(),
@@ -68,6 +80,18 @@ const CoderAgentsPage: FC = () => {
 		enabled: canEditDeploymentConfig,
 	});
 	const modelConfigsQuery = useQuery(chatModelConfigs());
+	const advisorConfigQuery = useQuery({
+		...chatAdvisorConfig(),
+		enabled: canEditDeploymentConfig && showAdvisorSettings,
+	});
+	const desktopEnabledQuery = useQuery({
+		...chatDesktopEnabled(),
+		enabled: canEditDeploymentConfig && showVirtualDesktopSettings,
+	});
+	const computerUseProviderQuery = useQuery({
+		...chatComputerUseProvider(),
+		enabled: canEditDeploymentConfig && showVirtualDesktopSettings,
+	});
 	const savePersonalModelOverridesAdminSettingsMutation = useMutation(
 		updateChatPersonalModelOverridesAdminSettings(queryClient),
 	);
@@ -82,6 +106,15 @@ const CoderAgentsPage: FC = () => {
 	);
 	const saveExploreModelOverrideMutation = useMutation(
 		updateChatModelOverrideMutation(queryClient, exploreOverrideContext),
+	);
+	const saveAdvisorConfigMutation = useMutation(
+		updateChatAdvisorConfig(queryClient),
+	);
+	const saveDesktopEnabledMutation = useMutation(
+		updateChatDesktopEnabled(queryClient),
+	);
+	const saveComputerUseProviderMutation = useMutation(
+		updateChatComputerUseProvider(queryClient),
 	);
 
 	return (
@@ -111,6 +144,7 @@ const CoderAgentsPage: FC = () => {
 				modelConfigsData={modelConfigsQuery.data}
 				modelConfigsError={modelConfigsQuery.error}
 				isLoadingModelConfigs={modelConfigsQuery.isLoading}
+				isFetchingModelConfigs={modelConfigsQuery.isFetching}
 				onSaveGeneralModelOverride={saveGeneralModelOverrideMutation.mutate}
 				isSavingGeneralModelOverride={
 					saveGeneralModelOverrideMutation.isPending
@@ -132,6 +166,26 @@ const CoderAgentsPage: FC = () => {
 				isSaveExploreModelOverrideError={
 					saveExploreModelOverrideMutation.isError
 				}
+				showAdvisorSettings={showAdvisorSettings}
+				advisorConfigData={advisorConfigQuery.data}
+				isAdvisorConfigLoading={advisorConfigQuery.isLoading}
+				isAdvisorConfigFetching={advisorConfigQuery.isFetching}
+				isAdvisorConfigLoadError={advisorConfigQuery.isError}
+				onSaveAdvisorConfig={saveAdvisorConfigMutation.mutate}
+				isSavingAdvisorConfig={saveAdvisorConfigMutation.isPending}
+				isSaveAdvisorConfigError={saveAdvisorConfigMutation.isError}
+				saveAdvisorConfigError={saveAdvisorConfigMutation.error}
+				showVirtualDesktopSettings={showVirtualDesktopSettings}
+				desktopEnabledData={desktopEnabledQuery.data}
+				isLoadingDesktopEnabled={desktopEnabledQuery.isLoading}
+				onSaveDesktopEnabled={saveDesktopEnabledMutation.mutate}
+				isSavingDesktopEnabled={saveDesktopEnabledMutation.isPending}
+				isSaveDesktopEnabledError={saveDesktopEnabledMutation.isError}
+				computerUseProviderData={computerUseProviderQuery.data}
+				isLoadingComputerUseProvider={computerUseProviderQuery.isLoading}
+				onSaveComputerUseProvider={saveComputerUseProviderMutation.mutate}
+				isSavingComputerUseProvider={saveComputerUseProviderMutation.isPending}
+				computerUseProviderSaveError={saveComputerUseProviderMutation.error}
 			/>
 		</RequirePermission>
 	);
