+++ site/src/pages/AISettingsPage/CoderAgentsPage/CoderAgentsPageView.stories.tsx
@@ -120,6 +120,7 @@ const buildArgs = (
 	modelConfigsData: allModelConfigs,
 	modelConfigsError: undefined,
 	isLoadingModelConfigs: false,
+	isFetchingModelConfigs: false,
 	onSaveGeneralModelOverride: fn(),
 	isSavingGeneralModelOverride: false,
 	isSaveGeneralModelOverrideError: false,
@@ -129,6 +130,26 @@ const buildArgs = (
 	onSaveExploreModelOverride: fn(),
 	isSavingExploreModelOverride: false,
 	isSaveExploreModelOverrideError: false,
+	showAdvisorSettings: false,
+	advisorConfigData: undefined,
+	isAdvisorConfigLoading: false,
+	isAdvisorConfigFetching: false,
+	isAdvisorConfigLoadError: false,
+	onSaveAdvisorConfig: fn(),
+	isSavingAdvisorConfig: false,
+	isSaveAdvisorConfigError: false,
+	saveAdvisorConfigError: undefined,
+	showVirtualDesktopSettings: false,
+	desktopEnabledData: undefined,
+	isLoadingDesktopEnabled: false,
+	onSaveDesktopEnabled: fn(),
+	isSavingDesktopEnabled: false,
+	isSaveDesktopEnabledError: false,
+	computerUseProviderData: undefined,
+	isLoadingComputerUseProvider: false,
+	onSaveComputerUseProvider: fn(),
+	isSavingComputerUseProvider: false,
+	computerUseProviderSaveError: null,
 	...overrides,
 });
