+++ site/src/pages/AgentsPage/components/ChatsSidebar/settings/SettingsPanel.tsx
@@ -4,7 +4,6 @@ import {
 	BotIcon,
 	BoxesIcon,
 	CoinsIcon,
-	FlaskConicalIcon,
 	KeyIcon,
 	LayoutTemplateIcon,
 	PanelLeftCloseIcon,
@@ -202,13 +201,6 @@ export const SettingsPanel: FC<SettingsPanelProps> = ({
 						active={false}
 						to="/ai/settings/instructions"
 						trailingIcon={ArrowUpRightIcon}
-					/>
-					<SettingsNavItem
-						icon={FlaskConicalIcon}
-						label="Experiments"
-						active={settingsSection === "experiments"}
-						to="/agents/settings/experiments"
-						state={location.state}
 					/>
 					<SettingsNavItem
 						icon={RefreshCwIcon}
